car_security_project_offline/car_security/requirements.txt
# requirements.txt
# Install with: pip install -r requirements.txt --break-system-packages

# Core computer vision & face recognition
opencv-python==4.9.0.80
face-recognition==1.3.0
dlib==19.24.4

# Deep learning (CPU-only build for RPi4)
torch==2.2.2
torchaudio==2.2.2

# Speaker verification
speechbrain==1.0.0

# Audio recording
pyaudio==0.2.14

# Serial communication (RPi ↔ ESP32)
pyserial==3.5

# Raspberry Pi GPIO
RPi.GPIO==0.7.1

# Utilities
numpy==1.26.4
